+++ java/client/src/main/java/glide/api/GlideClient.java
@@ -1,8 +1,23 @@
 /** Copyright Valkey GLIDE Project Contributors - SPDX Identifier: Apache-2.0 */
 package glide.api;
 
+import static command_request.CommandRequestOuterClass.RequestType.Auth;
+import static command_request.CommandRequestOuterClass.RequestType.ClientCaching;
 import static command_request.CommandRequestOuterClass.RequestType.ClientGetName;
+import static command_request.CommandRequestOuterClass.RequestType.ClientGetRedir;
 import static command_request.CommandRequestOuterClass.RequestType.ClientId;
+import static command_request.CommandRequestOuterClass.RequestType.ClientInfo;
+import static command_request.CommandRequestOuterClass.RequestType.ClientKillSimple;
+import static command_request.CommandRequestOuterClass.RequestType.ClientList;
+import static command_request.CommandRequestOuterClass.RequestType.ClientNoEvict;
+import static command_request.CommandRequestOuterClass.RequestType.ClientNoTouch;
+import static command_request.CommandRequestOuterClass.RequestType.ClientPause;
+import static command_request.CommandRequestOuterClass.RequestType.ClientReply;
+import static command_request.CommandRequestOuterClass.RequestType.ClientSetInfo;
+import static command_request.CommandRequestOuterClass.RequestType.ClientSetName;
+import static command_request.CommandRequestOuterClass.RequestType.ClientTrackingInfo;
+import static command_request.CommandRequestOuterClass.RequestType.ClientUnblock;
+import static command_request.CommandRequestOuterClass.RequestType.ClientUnpause;
 import static command_request.CommandRequestOuterClass.RequestType.ConfigGet;
 import static command_request.CommandRequestOuterClass.RequestType.ConfigResetStat;
 import static command_request.CommandRequestOuterClass.RequestType.ConfigRewrite;
@@ -20,11 +35,14 @@ import static command_request.CommandRequestOuterClass.RequestType.FunctionList;
 import static command_request.CommandRequestOuterClass.RequestType.FunctionLoad;
 import static command_request.CommandRequestOuterClass.RequestType.FunctionRestore;
 import static command_request.CommandRequestOuterClass.RequestType.FunctionStats;
+import static command_request.CommandRequestOuterClass.RequestType.Hello;
 import static command_request.CommandRequestOuterClass.RequestType.Info;
 import static command_request.CommandRequestOuterClass.RequestType.LastSave;
 import static command_request.CommandRequestOuterClass.RequestType.Lolwut;
 import static command_request.CommandRequestOuterClass.RequestType.Ping;
+import static command_request.CommandRequestOuterClass.RequestType.Quit;
 import static command_request.CommandRequestOuterClass.RequestType.RandomKey;
+import static command_request.CommandRequestOuterClass.RequestType.Reset;
 import static command_request.CommandRequestOuterClass.RequestType.Scan;
 import static command_request.CommandRequestOuterClass.RequestType.Select;
 import static command_request.CommandRequestOuterClass.RequestType.Time;
@@ -45,6 +63,7 @@ import glide.api.commands.TransactionsCommands;
 import glide.api.models.Batch;
 import glide.api.models.GlideString;
 import glide.api.models.Transaction;
+import glide.api.models.commands.ClientReplyMode;
 import glide.api.models.commands.FlushMode;
 import glide.api.models.commands.InfoOptions.Section;
 import glide.api.models.commands.batch.BatchOptions;
@@ -181,7 +200,7 @@ public class GlideClient extends BaseClient
 
     @Override
     public CompletableFuture<String> ping() {
-        return commandManager.submitNewCommand(Ping, new String[0], this::handleStringResponse);
+        return commandManager.submitNewCommand(Ping, EMPTY_STRING_ARRAY, this::handleStringResponse);
     }
 
     @Override
@@ -198,7 +217,7 @@ public class GlideClient extends BaseClient
 
     @Override
     public CompletableFuture<String> info() {
-        return commandManager.submitNewCommand(Info, new String[0], this::handleStringResponse);
+        return commandManager.submitNewCommand(Info, EMPTY_STRING_ARRAY, this::handleStringResponse);
     }
 
     @Override
@@ -216,26 +235,230 @@ public class GlideClient extends BaseClient
     }
 
     @Override
+    public CompletableFuture<String> auth(@NonNull String password) {
+        return commandManager.submitNewCommand(
+                Auth, new String[] {password}, this::handleStringResponse);
+    }
+
+    @Override
+    public CompletableFuture<String> auth(@NonNull String username, @NonNull String password) {
+        return commandManager.submitNewCommand(
+                Auth, new String[] {username, password}, this::handleStringResponse);
+    }
+
+    @Override
+    public CompletableFuture<String> auth(@NonNull GlideString password) {
+        return commandManager.submitNewCommand(
+                Auth, new GlideString[] {password}, this::handleStringResponse);
+    }
+
+    @Override
+    public CompletableFuture<String> auth(
+            @NonNull GlideString username, @NonNull GlideString password) {
+        return commandManager.submitNewCommand(
+                Auth, new GlideString[] {username, password}, this::handleStringResponse);
+    }
+
+    @Override
     public CompletableFuture<Long> clientId() {
-        return commandManager.submitNewCommand(ClientId, new String[0], this::handleLongResponse);
+        return commandManager.submitNewCommand(ClientId, EMPTY_STRING_ARRAY, this::handleLongResponse);
     }
 
     @Override
     public CompletableFuture<String> clientGetName() {
         return commandManager.submitNewCommand(
-                ClientGetName, new String[0], this::handleStringOrNullResponse);
+                ClientGetName, EMPTY_STRING_ARRAY, this::handleStringOrNullResponse);
+    }
+
+    @Override
+    public CompletableFuture<String> clientInfo() {
+        return commandManager.submitNewCommand(
+                ClientInfo, EMPTY_STRING_ARRAY, this::handleStringResponse);
+    }
+
+    @Override
+    public CompletableFuture<String> clientKillSimple(@NonNull String ipPort) {
+        return commandManager.submitNewCommand(
+                ClientKillSimple, new String[] {ipPort}, this::handleStringResponse);
+    }
+
+    @Override
+    public CompletableFuture<String> clientKillSimple(@NonNull GlideString ipPort) {
+        return commandManager.submitNewCommand(
+                ClientKillSimple, new GlideString[] {ipPort}, this::handleStringResponse);
+    }
+
+    @Override
+    public CompletableFuture<String> clientList() {
+        return commandManager.submitNewCommand(
+                ClientList, EMPTY_STRING_ARRAY, this::handleStringResponse);
+    }
+
+    @Override
+    public CompletableFuture<String> clientNoEvict(boolean enabled) {
+        return commandManager.submitNewCommand(
+                ClientNoEvict, new String[] {enabled ? "ON" : "OFF"}, this::handleStringResponse);
+    }
+
+    @Override
+    public CompletableFuture<String> clientNoTouch(boolean enabled) {
+        return commandManager.submitNewCommand(
+                ClientNoTouch, new String[] {enabled ? "ON" : "OFF"}, this::handleStringResponse);
+    }
+
+    @Override
+    public CompletableFuture<String> clientPause(long timeout) {
+        return commandManager.submitNewCommand(
+                ClientPause, new String[] {Long.toString(timeout)}, this::handleStringResponse);
+    }
+
+    @Override
+    public CompletableFuture<String> clientUnpause() {
+        return commandManager.submitNewCommand(
+                ClientUnpause, EMPTY_STRING_ARRAY, this::handleStringResponse);
+    }
+
+    @Override
+    public CompletableFuture<String> clientSetName(@NonNull String connectionName) {
+        return commandManager.submitNewCommand(
+                ClientSetName, new String[] {connectionName}, this::handleStringResponse);
+    }
+
+    @Override
+    public CompletableFuture<String> clientSetName(@NonNull GlideString connectionName) {
+        return commandManager.submitNewCommand(
+                ClientSetName, new GlideString[] {connectionName}, this::handleStringResponse);
+    }
+
+    @Override
+    public CompletableFuture<Long> clientUnblock(long clientId) {
+        return commandManager.submitNewCommand(
+                ClientUnblock, new String[] {Long.toString(clientId)}, this::handleLongResponse);
+    }
+
+    @Override
+    public CompletableFuture<Long> clientUnblock(long clientId, boolean withError) {
+        return commandManager.submitNewCommand(
+                ClientUnblock,
+                withError
+                        ? new String[] {Long.toString(clientId), "ERROR"}
+                        : new String[] {Long.toString(clientId)},
+                this::handleLongResponse);
+    }
+
+    @Override
+    public CompletableFuture<Long> clientGetRedir() {
+        return commandManager.submitNewCommand(
+                ClientGetRedir, EMPTY_STRING_ARRAY, this::handleLongResponse);
+    }
+
+    @Override
+    public CompletableFuture<Object[]> clientTrackingInfo() {
+        return commandManager.submitNewCommand(
+                ClientTrackingInfo, EMPTY_STRING_ARRAY, this::handleArrayResponse);
+    }
+
+    @Override
+    public CompletableFuture<String> clientCaching(boolean enabled) {
+        return commandManager.submitNewCommand(
+                ClientCaching, new String[] {enabled ? "YES" : "NO"}, this::handleStringResponse);
+    }
+
+    @Override
+    public CompletableFuture<String> clientSetInfo(@NonNull String attribute, @NonNull String value) {
+        return commandManager.submitNewCommand(
+                ClientSetInfo, new String[] {attribute, value}, this::handleStringResponse);
+    }
+
+    @Override
+    public CompletableFuture<String> clientSetInfo(
+            @NonNull GlideString attribute, @NonNull GlideString value) {
+        return commandManager.submitNewCommand(
+                ClientSetInfo, new GlideString[] {attribute, value}, this::handleStringResponse);
+    }
+
+    @Override
+    public CompletableFuture<String> clientReply(@NonNull ClientReplyMode mode) {
+        return commandManager.submitNewCommand(ClientReply, mode.toArgs(), this::handleStringResponse);
+    }
+
+    @Override
+    public CompletableFuture<String> quit() {
+        return commandManager.submitNewCommand(Quit, EMPTY_STRING_ARRAY, this::handleStringResponse);
+    }
+
+    @Override
+    public CompletableFuture<String> reset() {
+        return commandManager.submitNewCommand(Reset, EMPTY_STRING_ARRAY, this::handleStringResponse);
+    }
+
+    @Override
+    public CompletableFuture<Map<String, Object>> hello(long protocolVersion) {
+        return commandManager.submitNewCommand(
+                Hello, new String[] {Long.toString(protocolVersion)}, this::handleMapResponse);
+    }
+
+    @Override
+    public CompletableFuture<Map<String, Object>> hello(
+            long protocolVersion, @NonNull String username, @NonNull String password) {
+        return commandManager.submitNewCommand(
+                Hello,
+                new String[] {Long.toString(protocolVersion), "AUTH", username, password},
+                this::handleMapResponse);
+    }
+
+    @Override
+    public CompletableFuture<Map<String, Object>> hello(
+            long protocolVersion, @NonNull GlideString username, @NonNull GlideString password) {
+        return commandManager.submitNewCommand(
+                Hello,
+                new GlideString[] {gs(Long.toString(protocolVersion)), gs("AUTH"), username, password},
+                this::handleMapResponse);
+    }
+
+    @Override
+    public CompletableFuture<Map<String, Object>> hello(
+            long protocolVersion,
+            @NonNull String username,
+            @NonNull String password,
+            @NonNull String clientName) {
+        return commandManager.submitNewCommand(
+                Hello,
+                new String[] {
+                    Long.toString(protocolVersion), "AUTH", username, password, "SETNAME", clientName
+                },
+                this::handleMapResponse);
+    }
+
+    @Override
+    public CompletableFuture<Map<String, Object>> hello(
+            long protocolVersion,
+            @NonNull GlideString username,
+            @NonNull GlideString password,
+            @NonNull GlideString clientName) {
+        return commandManager.submitNewCommand(
+                Hello,
+                new GlideString[] {
+                    gs(Long.toString(protocolVersion)),
+                    gs("AUTH"),
+                    username,
+                    password,
+                    gs("SETNAME"),
+                    clientName
+                },
+                this::handleMapResponse);
     }
 
     @Override
     public CompletableFuture<String> configRewrite() {
         return commandManager.submitNewCommand(
-                ConfigRewrite, new String[0], this::handleStringResponse);
+                ConfigRewrite, EMPTY_STRING_ARRAY, this::handleStringResponse);
     }
 
     @Override
     public CompletableFuture<String> configResetStat() {
         return commandManager.submitNewCommand(
-                ConfigResetStat, new String[0], this::handleStringResponse);
+                ConfigResetStat, EMPTY_STRING_ARRAY, this::handleStringResponse);
     }
 
     @Override
@@ -264,17 +487,20 @@ public class GlideClient extends BaseClient
     @Override
     public CompletableFuture<String[]> time() {
         return commandManager.submitNewCommand(
-                Time, new String[0], response -> castArray(handleArrayResponse(response), String.class));
+                Time,
+                EMPTY_STRING_ARRAY,
+                response -> castArray(handleArrayResponse(response), String.class));
     }
 
     @Override
     public CompletableFuture<Long> lastsave() {
-        return commandManager.submitNewCommand(LastSave, new String[0], this::handleLongResponse);
+        return commandManager.submitNewCommand(LastSave, EMPTY_STRING_ARRAY, this::handleLongResponse);
     }
 
     @Override
     public CompletableFuture<String> flushall() {
-        return commandManager.submitNewCommand(FlushAll, new String[0], this::handleStringResponse);
+        return commandManager.submitNewCommand(
+                FlushAll, EMPTY_STRING_ARRAY, this::handleStringResponse);
     }
 
     @Override
@@ -285,7 +511,7 @@ public class GlideClient extends BaseClient
 
     @Override
     public CompletableFuture<String> flushdb() {
-        return commandManager.submitNewCommand(FlushDB, new String[0], this::handleStringResponse);
+        return commandManager.submitNewCommand(FlushDB, EMPTY_STRING_ARRAY, this::handleStringResponse);
     }
 
     @Override
@@ -296,7 +522,7 @@ public class GlideClient extends BaseClient
 
     @Override
     public CompletableFuture<String> lolwut() {
-        return commandManager.submitNewCommand(Lolwut, new String[0], this::handleStringResponse);
+        return commandManager.submitNewCommand(Lolwut, EMPTY_STRING_ARRAY, this::handleStringResponse);
     }
 
     @Override
@@ -325,7 +551,7 @@ public class GlideClient extends BaseClient
 
     @Override
     public CompletableFuture<Long> dbsize() {
-        return commandManager.submitNewCommand(DBSize, new String[0], this::handleLongResponse);
+        return commandManager.submitNewCommand(DBSize, EMPTY_STRING_ARRAY, this::handleLongResponse);
     }
 
     @Override
@@ -350,7 +576,7 @@ public class GlideClient extends BaseClient
     public CompletableFuture<Map<String, Object>[]> functionList(boolean withCode) {
         return commandManager.submitNewCommand(
                 FunctionList,
-                withCode ? new String[] {WITH_CODE_VALKEY_API} : new String[0],
+                withCode ? new String[] {WITH_CODE_VALKEY_API} : EMPTY_STRING_ARRAY,
                 response -> handleFunctionListResponse(handleArrayResponse(response)));
     }
 
@@ -389,7 +615,7 @@ public class GlideClient extends BaseClient
     @Override
     public CompletableFuture<String> functionFlush() {
         return commandManager.submitNewCommand(
-                FunctionFlush, new String[0], this::handleStringResponse);
+                FunctionFlush, EMPTY_STRING_ARRAY, this::handleStringResponse);
     }
 
     @Override
@@ -433,7 +659,7 @@ public class GlideClient extends BaseClient
 
     @Override
     public CompletableFuture<Object> fcall(@NonNull String function) {
-        return fcall(function, new String[0], new String[0]);
+        return fcall(function, EMPTY_STRING_ARRAY, EMPTY_STRING_ARRAY);
     }
 
     @Override
@@ -444,7 +670,7 @@ public class GlideClient extends BaseClient
 
     @Override
     public CompletableFuture<Object> fcallReadOnly(@NonNull String function) {
-        return fcallReadOnly(function, new String[0], new String[0]);
+        return fcallReadOnly(function, EMPTY_STRING_ARRAY, EMPTY_STRING_ARRAY);
     }
 
     @Override
@@ -455,14 +681,15 @@ public class GlideClient extends BaseClient
 
     @Override
     public CompletableFuture<String> functionKill() {
-        return commandManager.submitNewCommand(FunctionKill, new String[0], this::handleStringResponse);
+        return commandManager.submitNewCommand(
+                FunctionKill, EMPTY_STRING_ARRAY, this::handleStringResponse);
     }
 
     @Override
     public CompletableFuture<Map<String, Map<String, Map<String, Object>>>> functionStats() {
         return commandManager.submitNewCommand(
                 FunctionStats,
-                new String[0],
+                EMPTY_STRING_ARRAY,
                 response -> handleFunctionStatsResponse(response, false).getMultiValue());
     }
 
@@ -477,13 +704,13 @@ public class GlideClient extends BaseClient
 
     @Override
     public CompletableFuture<String> unwatch() {
-        return commandManager.submitNewCommand(UnWatch, new String[0], this::handleStringResponse);
+        return commandManager.submitNewCommand(UnWatch, EMPTY_STRING_ARRAY, this::handleStringResponse);
     }
 
     @Override
     public CompletableFuture<String> randomKey() {
         return commandManager.submitNewCommand(
-                RandomKey, new String[0], this::handleStringOrNullResponse);
+                RandomKey, EMPTY_STRING_ARRAY, this::handleStringOrNullResponse);
     }
 
     @Override