+++ java/client/src/main/java/glide/api/GlideClusterClient.java
@@ -1,8 +1,23 @@
 /** Copyright Valkey GLIDE Project Contributors - SPDX Identifier: Apache-2.0 */
 package glide.api;
 
+import static command_request.CommandRequestOuterClass.RequestType.Auth;
+import static command_request.CommandRequestOuterClass.RequestType.ClientCaching;
 import static command_request.CommandRequestOuterClass.RequestType.ClientGetName;
+import static command_request.CommandRequestOuterClass.RequestType.ClientGetRedir;
 import static command_request.CommandRequestOuterClass.RequestType.ClientId;
+import static command_request.CommandRequestOuterClass.RequestType.ClientInfo;
+import static command_request.CommandRequestOuterClass.RequestType.ClientKillSimple;
+import static command_request.CommandRequestOuterClass.RequestType.ClientList;
+import static command_request.CommandRequestOuterClass.RequestType.ClientNoEvict;
+import static command_request.CommandRequestOuterClass.RequestType.ClientNoTouch;
+import static command_request.CommandRequestOuterClass.RequestType.ClientPause;
+import static command_request.CommandRequestOuterClass.RequestType.ClientReply;
+import static command_request.CommandRequestOuterClass.RequestType.ClientSetInfo;
+import static command_request.CommandRequestOuterClass.RequestType.ClientSetName;
+import static command_request.CommandRequestOuterClass.RequestType.ClientTrackingInfo;
+import static command_request.CommandRequestOuterClass.RequestType.ClientUnblock;
+import static command_request.CommandRequestOuterClass.RequestType.ClientUnpause;
 import static command_request.CommandRequestOuterClass.RequestType.ConfigGet;
 import static command_request.CommandRequestOuterClass.RequestType.ConfigResetStat;
 import static command_request.CommandRequestOuterClass.RequestType.ConfigRewrite;
@@ -22,13 +37,16 @@ import static command_request.CommandRequestOuterClass.RequestType.FunctionList;
 import static command_request.CommandRequestOuterClass.RequestType.FunctionLoad;
 import static command_request.CommandRequestOuterClass.RequestType.FunctionRestore;
 import static command_request.CommandRequestOuterClass.RequestType.FunctionStats;
+import static command_request.CommandRequestOuterClass.RequestType.Hello;
 import static command_request.CommandRequestOuterClass.RequestType.Info;
 import static command_request.CommandRequestOuterClass.RequestType.LastSave;
 import static command_request.CommandRequestOuterClass.RequestType.Lolwut;
 import static command_request.CommandRequestOuterClass.RequestType.Ping;
 import static command_request.CommandRequestOuterClass.RequestType.PubSubShardChannels;
 import static command_request.CommandRequestOuterClass.RequestType.PubSubShardNumSub;
+import static command_request.CommandRequestOuterClass.RequestType.Quit;
 import static command_request.CommandRequestOuterClass.RequestType.RandomKey;
+import static command_request.CommandRequestOuterClass.RequestType.Reset;
 import static command_request.CommandRequestOuterClass.RequestType.SPublish;
 import static command_request.CommandRequestOuterClass.RequestType.ScriptExists;
 import static command_request.CommandRequestOuterClass.RequestType.ScriptFlush;
@@ -59,6 +77,7 @@ import glide.api.models.ClusterTransaction;
 import glide.api.models.ClusterValue;
 import glide.api.models.GlideString;
 import glide.api.models.Script;
+import glide.api.models.commands.ClientReplyMode;
 import glide.api.models.commands.FlushMode;
 import glide.api.models.commands.InfoOptions.Section;
 import glide.api.models.commands.ScriptArgOptions;
@@ -274,7 +293,7 @@ public class GlideClusterClient extends BaseClient
 
     @Override
     public CompletableFuture<String> ping() {
-        return commandManager.submitNewCommand(Ping, new String[0], this::handleStringResponse);
+        return commandManager.submitNewCommand(Ping, EMPTY_STRING_ARRAY, this::handleStringResponse);
     }
 
     @Override
@@ -291,7 +310,8 @@ public class GlideClusterClient extends BaseClient
 
     @Override
     public CompletableFuture<String> ping(@NonNull Route route) {
-        return commandManager.submitNewCommand(Ping, new String[0], route, this::handleStringResponse);
+        return commandManager.submitNewCommand(
+                Ping, EMPTY_STRING_ARRAY, route, this::handleStringResponse);
     }
 
     @Override
@@ -309,13 +329,13 @@ public class GlideClusterClient extends BaseClient
     @Override
     public CompletableFuture<ClusterValue<String>> info() {
         return commandManager.submitNewCommand(
-                Info, new String[0], response -> ClusterValue.of(handleMapResponse(response)));
+                Info, EMPTY_STRING_ARRAY, response -> ClusterValue.of(handleMapResponse(response)));
     }
 
     public CompletableFuture<ClusterValue<String>> info(@NonNull Route route) {
         return commandManager.submitNewCommand(
                 Info,
-                new String[0],
+                EMPTY_STRING_ARRAY,
                 route,
                 response ->
                         route instanceof SingleNodeRoute
@@ -352,14 +372,14 @@ public class GlideClusterClient extends BaseClient
 
     @Override
     public CompletableFuture<Long> clientId() {
-        return commandManager.submitNewCommand(ClientId, new String[0], this::handleLongResponse);
+        return commandManager.submitNewCommand(ClientId, EMPTY_STRING_ARRAY, this::handleLongResponse);
     }
 
     @Override
     public CompletableFuture<ClusterValue<Long>> clientId(@NonNull Route route) {
         return commandManager.submitNewCommand(
                 ClientId,
-                new String[0],
+                EMPTY_STRING_ARRAY,
                 route,
                 response ->
                         route instanceof SingleNodeRoute
@@ -370,14 +390,14 @@ public class GlideClusterClient extends BaseClient
     @Override
     public CompletableFuture<String> clientGetName() {
         return commandManager.submitNewCommand(
-                ClientGetName, new String[0], this::handleStringOrNullResponse);
+                ClientGetName, EMPTY_STRING_ARRAY, this::handleStringOrNullResponse);
     }
 
     @Override
     public CompletableFuture<ClusterValue<String>> clientGetName(@NonNull Route route) {
         return commandManager.submitNewCommand(
                 ClientGetName,
-                new String[0],
+                EMPTY_STRING_ARRAY,
                 route,
                 response ->
                         route instanceof SingleNodeRoute
@@ -386,27 +406,355 @@ public class GlideClusterClient extends BaseClient
     }
 
     @Override
+    public CompletableFuture<String> auth(@NonNull String password) {
+        return commandManager.submitNewCommand(
+                Auth, new String[] {password}, this::handleStringResponse);
+    }
+
+    @Override
+    public CompletableFuture<String> auth(@NonNull String username, @NonNull String password) {
+        return commandManager.submitNewCommand(
+                Auth, new String[] {username, password}, this::handleStringResponse);
+    }
+
+    @Override
+    public CompletableFuture<String> auth(@NonNull GlideString password) {
+        return commandManager.submitNewCommand(
+                Auth, new GlideString[] {password}, this::handleStringResponse);
+    }
+
+    @Override
+    public CompletableFuture<String> auth(
+            @NonNull GlideString username, @NonNull GlideString password) {
+        return commandManager.submitNewCommand(
+                Auth, new GlideString[] {username, password}, this::handleStringResponse);
+    }
+
+    @Override
+    public CompletableFuture<String> clientInfo() {
+        return commandManager.submitNewCommand(
+                ClientInfo, EMPTY_STRING_ARRAY, this::handleStringResponse);
+    }
+
+    @Override
+    public CompletableFuture<ClusterValue<String>> clientInfo(@NonNull Route route) {
+        return commandManager.submitNewCommand(
+                ClientInfo,
+                EMPTY_STRING_ARRAY,
+                route,
+                response ->
+                        route instanceof SingleNodeRoute
+                                ? ClusterValue.of(handleStringResponse(response))
+                                : ClusterValue.of(handleMapResponse(response)));
+    }
+
+    @Override
+    public CompletableFuture<String> clientKillSimple(@NonNull String ipPort) {
+        return commandManager.submitNewCommand(
+                ClientKillSimple, new String[] {ipPort}, this::handleStringResponse);
+    }
+
+    @Override
+    public CompletableFuture<String> clientKillSimple(@NonNull GlideString ipPort) {
+        return commandManager.submitNewCommand(
+                ClientKillSimple, new GlideString[] {ipPort}, this::handleStringResponse);
+    }
+
+    @Override
+    public CompletableFuture<String> clientList() {
+        return commandManager.submitNewCommand(
+                ClientList, EMPTY_STRING_ARRAY, this::handleStringResponse);
+    }
+
+    @Override
+    public CompletableFuture<ClusterValue<String>> clientList(@NonNull Route route) {
+        return commandManager.submitNewCommand(
+                ClientList,
+                EMPTY_STRING_ARRAY,
+                route,
+                response ->
+                        route instanceof SingleNodeRoute
+                                ? ClusterValue.of(handleStringResponse(response))
+                                : ClusterValue.of(handleMapResponse(response)));
+    }
+
+    @Override
+    public CompletableFuture<String> clientNoEvict(boolean enabled) {
+        return commandManager.submitNewCommand(
+                ClientNoEvict, new String[] {enabled ? "ON" : "OFF"}, this::handleStringResponse);
+    }
+
+    @Override
+    public CompletableFuture<String> clientNoEvict(boolean enabled, @NonNull Route route) {
+        return commandManager.submitNewCommand(
+                ClientNoEvict, new String[] {enabled ? "ON" : "OFF"}, route, this::handleStringResponse);
+    }
+
+    @Override
+    public CompletableFuture<String> clientNoTouch(boolean enabled) {
+        return commandManager.submitNewCommand(
+                ClientNoTouch, new String[] {enabled ? "ON" : "OFF"}, this::handleStringResponse);
+    }
+
+    @Override
+    public CompletableFuture<String> clientNoTouch(boolean enabled, @NonNull Route route) {
+        return commandManager.submitNewCommand(
+                ClientNoTouch, new String[] {enabled ? "ON" : "OFF"}, route, this::handleStringResponse);
+    }
+
+    @Override
+    public CompletableFuture<String> clientPause(long timeout) {
+        return commandManager.submitNewCommand(
+                ClientPause, new String[] {Long.toString(timeout)}, this::handleStringResponse);
+    }
+
+    @Override
+    public CompletableFuture<String> clientPause(long timeout, @NonNull Route route) {
+        return commandManager.submitNewCommand(
+                ClientPause, new String[] {Long.toString(timeout)}, route, this::handleStringResponse);
+    }
+
+    @Override
+    public CompletableFuture<String> clientUnpause() {
+        return commandManager.submitNewCommand(
+                ClientUnpause, EMPTY_STRING_ARRAY, this::handleStringResponse);
+    }
+
+    @Override
+    public CompletableFuture<String> clientUnpause(@NonNull Route route) {
+        return commandManager.submitNewCommand(
+                ClientUnpause, EMPTY_STRING_ARRAY, route, this::handleStringResponse);
+    }
+
+    @Override
+    public CompletableFuture<String> clientSetName(@NonNull String connectionName) {
+        return commandManager.submitNewCommand(
+                ClientSetName, new String[] {connectionName}, this::handleStringResponse);
+    }
+
+    @Override
+    public CompletableFuture<String> clientSetName(@NonNull GlideString connectionName) {
+        return commandManager.submitNewCommand(
+                ClientSetName, new GlideString[] {connectionName}, this::handleStringResponse);
+    }
+
+    @Override
+    public CompletableFuture<Long> clientUnblock(long clientId) {
+        return commandManager.submitNewCommand(
+                ClientUnblock, new String[] {Long.toString(clientId)}, this::handleLongResponse);
+    }
+
+    @Override
+    public CompletableFuture<Long> clientUnblock(long clientId, boolean withError) {
+        return commandManager.submitNewCommand(
+                ClientUnblock,
+                withError
+                        ? new String[] {Long.toString(clientId), "ERROR"}
+                        : new String[] {Long.toString(clientId)},
+                this::handleLongResponse);
+    }
+
+    @Override
+    public CompletableFuture<ClusterValue<Long>> clientUnblock(long clientId, @NonNull Route route) {
+        return commandManager.submitNewCommand(
+                ClientUnblock,
+                new String[] {Long.toString(clientId)},
+                route,
+                response ->
+                        route instanceof SingleNodeRoute
+                                ? ClusterValue.of(handleLongResponse(response))
+                                : ClusterValue.of(handleMapResponse(response)));
+    }
+
+    @Override
+    public CompletableFuture<ClusterValue<Long>> clientUnblock(
+            long clientId, boolean withError, @NonNull Route route) {
+        return commandManager.submitNewCommand(
+                ClientUnblock,
+                withError
+                        ? new String[] {Long.toString(clientId), "ERROR"}
+                        : new String[] {Long.toString(clientId)},
+                route,
+                response ->
+                        route instanceof SingleNodeRoute
+                                ? ClusterValue.of(handleLongResponse(response))
+                                : ClusterValue.of(handleMapResponse(response)));
+    }
+
+    @Override
+    public CompletableFuture<Long> clientGetRedir() {
+        return commandManager.submitNewCommand(
+                ClientGetRedir, EMPTY_STRING_ARRAY, this::handleLongResponse);
+    }
+
+    @Override
+    public CompletableFuture<ClusterValue<Long>> clientGetRedir(@NonNull Route route) {
+        return commandManager.submitNewCommand(
+                ClientGetRedir,
+                EMPTY_STRING_ARRAY,
+                route,
+                response ->
+                        route instanceof SingleNodeRoute
+                                ? ClusterValue.of(handleLongResponse(response))
+                                : ClusterValue.of(handleMapResponse(response)));
+    }
+
+    @Override
+    public CompletableFuture<Object[]> clientTrackingInfo() {
+        return commandManager.submitNewCommand(
+                ClientTrackingInfo, EMPTY_STRING_ARRAY, this::handleArrayResponse);
+    }
+
+    @Override
+    public CompletableFuture<ClusterValue<Object[]>> clientTrackingInfo(@NonNull Route route) {
+        return commandManager.submitNewCommand(
+                ClientTrackingInfo,
+                EMPTY_STRING_ARRAY,
+                route,
+                response ->
+                        route instanceof SingleNodeRoute
+                                ? ClusterValue.of(handleArrayResponse(response))
+                                : ClusterValue.of(handleMapResponse(response)));
+    }
+
+    @Override
+    public CompletableFuture<String> clientCaching(boolean enabled) {
+        return commandManager.submitNewCommand(
+                ClientCaching, new String[] {enabled ? "YES" : "NO"}, this::handleStringResponse);
+    }
+
+    @Override
+    public CompletableFuture<String> clientCaching(boolean enabled, @NonNull Route route) {
+        return commandManager.submitNewCommand(
+                ClientCaching, new String[] {enabled ? "YES" : "NO"}, route, this::handleStringResponse);
+    }
+
+    @Override
+    public CompletableFuture<String> clientSetInfo(@NonNull String attribute, @NonNull String value) {
+        return commandManager.submitNewCommand(
+                ClientSetInfo, new String[] {attribute, value}, this::handleStringResponse);
+    }
+
+    @Override
+    public CompletableFuture<String> clientSetInfo(
+            @NonNull GlideString attribute, @NonNull GlideString value) {
+        return commandManager.submitNewCommand(
+                ClientSetInfo, new GlideString[] {attribute, value}, this::handleStringResponse);
+    }
+
+    @Override
+    public CompletableFuture<String> clientReply(@NonNull ClientReplyMode mode) {
+        return commandManager.submitNewCommand(ClientReply, mode.toArgs(), this::handleStringResponse);
+    }
+
+    @Override
+    public CompletableFuture<String> clientReply(
+            @NonNull ClientReplyMode mode, @NonNull Route route) {
+        return commandManager.submitNewCommand(
+                ClientReply, mode.toArgs(), route, this::handleStringResponse);
+    }
+
+    @Override
+    public CompletableFuture<String> quit() {
+        return commandManager.submitNewCommand(Quit, EMPTY_STRING_ARRAY, this::handleStringResponse);
+    }
+
+    @Override
+    public CompletableFuture<String> quit(@NonNull Route route) {
+        return commandManager.submitNewCommand(
+                Quit, EMPTY_STRING_ARRAY, route, this::handleStringResponse);
+    }
+
+    @Override
+    public CompletableFuture<String> reset() {
+        return commandManager.submitNewCommand(Reset, EMPTY_STRING_ARRAY, this::handleStringResponse);
+    }
+
+    @Override
+    public CompletableFuture<String> reset(@NonNull Route route) {
+        return commandManager.submitNewCommand(
+                Reset, EMPTY_STRING_ARRAY, route, this::handleStringResponse);
+    }
+
+    @Override
+    public CompletableFuture<Map<String, Object>> hello(long protocolVersion) {
+        return commandManager.submitNewCommand(
+                Hello, new String[] {Long.toString(protocolVersion)}, this::handleMapResponse);
+    }
+
+    @Override
+    public CompletableFuture<Map<String, Object>> hello(
+            long protocolVersion, @NonNull String username, @NonNull String password) {
+        return commandManager.submitNewCommand(
+                Hello,
+                new String[] {Long.toString(protocolVersion), "AUTH", username, password},
+                this::handleMapResponse);
+    }
+
+    @Override
+    public CompletableFuture<Map<String, Object>> hello(
+            long protocolVersion, @NonNull GlideString username, @NonNull GlideString password) {
+        return commandManager.submitNewCommand(
+                Hello,
+                new GlideString[] {gs(Long.toString(protocolVersion)), gs("AUTH"), username, password},
+                this::handleMapResponse);
+    }
+
+    @Override
+    public CompletableFuture<Map<String, Object>> hello(
+            long protocolVersion,
+            @NonNull String username,
+            @NonNull String password,
+            @NonNull String clientName) {
+        return commandManager.submitNewCommand(
+                Hello,
+                new String[] {
+                    Long.toString(protocolVersion), "AUTH", username, password, "SETNAME", clientName
+                },
+                this::handleMapResponse);
+    }
+
+    @Override
+    public CompletableFuture<Map<String, Object>> hello(
+            long protocolVersion,
+            @NonNull GlideString username,
+            @NonNull GlideString password,
+            @NonNull GlideString clientName) {
+        return commandManager.submitNewCommand(
+                Hello,
+                new GlideString[] {
+                    gs(Long.toString(protocolVersion)),
+                    gs("AUTH"),
+                    username,
+                    password,
+                    gs("SETNAME"),
+                    clientName
+                },
+                this::handleMapResponse);
+    }
+
+    @Override
     public CompletableFuture<String> configRewrite() {
         return commandManager.submitNewCommand(
-                ConfigRewrite, new String[0], this::handleStringResponse);
+                ConfigRewrite, EMPTY_STRING_ARRAY, this::handleStringResponse);
     }
 
     @Override
     public CompletableFuture<String> configRewrite(@NonNull Route route) {
         return commandManager.submitNewCommand(
-                ConfigRewrite, new String[0], route, this::handleStringResponse);
+                ConfigRewrite, EMPTY_STRING_ARRAY, route, this::handleStringResponse);
     }
 
     @Override
     public CompletableFuture<String> configResetStat() {
         return commandManager.submitNewCommand(
-                ConfigResetStat, new String[0], this::handleStringResponse);
+                ConfigResetStat, EMPTY_STRING_ARRAY, this::handleStringResponse);
     }
 
     @Override
     public CompletableFuture<String> configResetStat(@NonNull Route route) {
         return commandManager.submitNewCommand(
-                ConfigResetStat, new String[0], route, this::handleStringResponse);
+                ConfigResetStat, EMPTY_STRING_ARRAY, route, this::handleStringResponse);
     }
 
     @Override
@@ -481,14 +829,16 @@ public class GlideClusterClient extends BaseClient
     @Override
     public CompletableFuture<String[]> time() {
         return commandManager.submitNewCommand(
-                Time, new String[0], response -> castArray(handleArrayResponse(response), String.class));
+                Time,
+                EMPTY_STRING_ARRAY,
+                response -> castArray(handleArrayResponse(response), String.class));
     }
 
     @Override
     public CompletableFuture<ClusterValue<String[]>> time(@NonNull Route route) {
         return commandManager.submitNewCommand(
                 Time,
-                new String[0],
+                EMPTY_STRING_ARRAY,
                 route,
                 response ->
                         route instanceof SingleNodeRoute
@@ -499,14 +849,14 @@ public class GlideClusterClient extends BaseClient
 
     @Override
     public CompletableFuture<Long> lastsave() {
-        return commandManager.submitNewCommand(LastSave, new String[0], this::handleLongResponse);
+        return commandManager.submitNewCommand(LastSave, EMPTY_STRING_ARRAY, this::handleLongResponse);
     }
 
     @Override
     public CompletableFuture<ClusterValue<Long>> lastsave(@NonNull Route route) {
         return commandManager.submitNewCommand(
                 LastSave,
-                new String[0],
+                EMPTY_STRING_ARRAY,
                 route,
                 response ->
                         route instanceof SingleNodeRoute
@@ -516,7 +866,8 @@ public class GlideClusterClient extends BaseClient
 
     @Override
     public CompletableFuture<String> flushall() {
-        return commandManager.submitNewCommand(FlushAll, new String[0], this::handleStringResponse);
+        return commandManager.submitNewCommand(
+                FlushAll, EMPTY_STRING_ARRAY, this::handleStringResponse);
     }
 
     @Override
@@ -528,7 +879,7 @@ public class GlideClusterClient extends BaseClient
     @Override
     public CompletableFuture<String> flushall(@NonNull Route route) {
         return commandManager.submitNewCommand(
-                FlushAll, new String[0], route, this::handleStringResponse);
+                FlushAll, EMPTY_STRING_ARRAY, route, this::handleStringResponse);
     }
 
     @Override
@@ -539,7 +890,7 @@ public class GlideClusterClient extends BaseClient
 
     @Override
     public CompletableFuture<String> flushdb() {
-        return commandManager.submitNewCommand(FlushDB, new String[0], this::handleStringResponse);
+        return commandManager.submitNewCommand(FlushDB, EMPTY_STRING_ARRAY, this::handleStringResponse);
     }
 
     @Override
@@ -551,7 +902,7 @@ public class GlideClusterClient extends BaseClient
     @Override
     public CompletableFuture<String> flushdb(@NonNull Route route) {
         return commandManager.submitNewCommand(
-                FlushDB, new String[0], route, this::handleStringResponse);
+                FlushDB, EMPTY_STRING_ARRAY, route, this::handleStringResponse);
     }
 
     @Override
@@ -562,7 +913,7 @@ public class GlideClusterClient extends BaseClient
 
     @Override
     public CompletableFuture<String> lolwut() {
-        return commandManager.submitNewCommand(Lolwut, new String[0], this::handleStringResponse);
+        return commandManager.submitNewCommand(Lolwut, EMPTY_STRING_ARRAY, this::handleStringResponse);
     }
 
     @Override
@@ -593,7 +944,7 @@ public class GlideClusterClient extends BaseClient
     public CompletableFuture<ClusterValue<String>> lolwut(@NonNull Route route) {
         return commandManager.submitNewCommand(
                 Lolwut,
-                new String[0],
+                EMPTY_STRING_ARRAY,
                 route,
                 response ->
                         route instanceof SingleNodeRoute
@@ -647,12 +998,13 @@ public class GlideClusterClient extends BaseClient
 
     @Override
     public CompletableFuture<Long> dbsize() {
-        return commandManager.submitNewCommand(DBSize, new String[0], this::handleLongResponse);
+        return commandManager.submitNewCommand(DBSize, EMPTY_STRING_ARRAY, this::handleLongResponse);
     }
 
     @Override
     public CompletableFuture<Long> dbsize(@NonNull Route route) {
-        return commandManager.submitNewCommand(DBSize, new String[0], route, this::handleLongResponse);
+        return commandManager.submitNewCommand(
+                DBSize, EMPTY_STRING_ARRAY, route, this::handleLongResponse);
     }
 
     @Override
@@ -733,7 +1085,7 @@ public class GlideClusterClient extends BaseClient
     public CompletableFuture<Map<String, Object>[]> functionList(boolean withCode) {
         return commandManager.submitNewCommand(
                 FunctionList,
-                withCode ? new String[] {WITH_CODE_VALKEY_API} : new String[0],
+                withCode ? new String[] {WITH_CODE_VALKEY_API} : EMPTY_STRING_ARRAY,
                 response -> handleFunctionListResponse(handleArrayResponse(response)));
     }
 
@@ -774,7 +1126,7 @@ public class GlideClusterClient extends BaseClient
             boolean withCode, @NonNull Route route) {
         return commandManager.submitNewCommand(
                 FunctionList,
-                withCode ? new String[] {WITH_CODE_VALKEY_API} : new String[0],
+                withCode ? new String[] {WITH_CODE_VALKEY_API} : EMPTY_STRING_ARRAY,
                 route,
                 response -> handleFunctionListResponse(response, route));
     }
@@ -816,7 +1168,7 @@ public class GlideClusterClient extends BaseClient
     @Override
     public CompletableFuture<String> functionFlush() {
         return commandManager.submitNewCommand(
-                FunctionFlush, new String[0], this::handleStringResponse);
+                FunctionFlush, EMPTY_STRING_ARRAY, this::handleStringResponse);
     }
 
     @Override
@@ -828,7 +1180,7 @@ public class GlideClusterClient extends BaseClient
     @Override
     public CompletableFuture<String> functionFlush(@NonNull Route route) {
         return commandManager.submitNewCommand(
-                FunctionFlush, new String[0], route, this::handleStringResponse);
+                FunctionFlush, EMPTY_STRING_ARRAY, route, this::handleStringResponse);
     }
 
     @Override
@@ -913,7 +1265,7 @@ public class GlideClusterClient extends BaseClient
 
     @Override
     public CompletableFuture<Object> fcall(@NonNull String function) {
-        return fcall(function, new String[0]);
+        return fcall(function, EMPTY_STRING_ARRAY);
     }
 
     @Override
@@ -924,7 +1276,7 @@ public class GlideClusterClient extends BaseClient
     @Override
     public CompletableFuture<ClusterValue<Object>> fcall(
             @NonNull String function, @NonNull Route route) {
-        return fcall(function, new String[0], route);
+        return fcall(function, EMPTY_STRING_ARRAY, route);
     }
 
     @Override
@@ -978,7 +1330,7 @@ public class GlideClusterClient extends BaseClient
 
     @Override
     public CompletableFuture<Object> fcallReadOnly(@NonNull String function) {
-        return fcallReadOnly(function, new String[0]);
+        return fcallReadOnly(function, EMPTY_STRING_ARRAY);
     }
 
     @Override
@@ -989,7 +1341,7 @@ public class GlideClusterClient extends BaseClient
     @Override
     public CompletableFuture<ClusterValue<Object>> fcallReadOnly(
             @NonNull String function, @NonNull Route route) {
-        return fcallReadOnly(function, new String[0], route);
+        return fcallReadOnly(function, EMPTY_STRING_ARRAY, route);
     }
 
     @Override
@@ -1045,13 +1397,14 @@ public class GlideClusterClient extends BaseClient
 
     @Override
     public CompletableFuture<String> functionKill() {
-        return commandManager.submitNewCommand(FunctionKill, new String[0], this::handleStringResponse);
+        return commandManager.submitNewCommand(
+                FunctionKill, EMPTY_STRING_ARRAY, this::handleStringResponse);
     }
 
     @Override
     public CompletableFuture<String> functionKill(@NonNull Route route) {
         return commandManager.submitNewCommand(
-                FunctionKill, new String[0], route, this::handleStringResponse);
+                FunctionKill, EMPTY_STRING_ARRAY, route, this::handleStringResponse);
     }
 
     @Override
@@ -1122,7 +1475,8 @@ public class GlideClusterClient extends BaseClient
 
     @Override
     public CompletableFuture<String> scriptFlush() {
-        return commandManager.submitNewCommand(ScriptFlush, new String[0], this::handleStringResponse);
+        return commandManager.submitNewCommand(
+                ScriptFlush, EMPTY_STRING_ARRAY, this::handleStringResponse);
     }
 
     @Override
@@ -1134,7 +1488,7 @@ public class GlideClusterClient extends BaseClient
     @Override
     public CompletableFuture<String> scriptFlush(@NonNull Route route) {
         return commandManager.submitNewCommand(
-                ScriptFlush, new String[0], route, this::handleStringResponse);
+                ScriptFlush, EMPTY_STRING_ARRAY, route, this::handleStringResponse);
     }
 
     @Override
@@ -1145,19 +1499,22 @@ public class GlideClusterClient extends BaseClient
 
     @Override
     public CompletableFuture<String> scriptKill() {
-        return commandManager.submitNewCommand(ScriptKill, new String[0], this::handleStringResponse);
+        return commandManager.submitNewCommand(
+                ScriptKill, EMPTY_STRING_ARRAY, this::handleStringResponse);
     }
 
     @Override
     public CompletableFuture<String> scriptKill(@NonNull Route route) {
         return commandManager.submitNewCommand(
-                ScriptKill, new String[0], route, this::handleStringResponse);
+                ScriptKill, EMPTY_STRING_ARRAY, route, this::handleStringResponse);
     }
 
     @Override
     public CompletableFuture<ClusterValue<Map<String, Map<String, Object>>>> functionStats() {
         return commandManager.submitNewCommand(
-                FunctionStats, new String[0], response -> handleFunctionStatsResponse(response, false));
+                FunctionStats,
+                EMPTY_STRING_ARRAY,
+                response -> handleFunctionStatsResponse(response, false));
     }
 
     @Override
@@ -1174,7 +1531,7 @@ public class GlideClusterClient extends BaseClient
             @NonNull Route route) {
         return commandManager.submitNewCommand(
                 FunctionStats,
-                new String[0],
+                EMPTY_STRING_ARRAY,
                 route,
                 response -> handleFunctionStatsResponse(response, route instanceof SingleNodeRoute));
     }
@@ -1226,7 +1583,7 @@ public class GlideClusterClient extends BaseClient
     public CompletableFuture<String[]> pubsubShardChannels() {
         return commandManager.submitNewCommand(
                 PubSubShardChannels,
-                new String[0],
+                EMPTY_STRING_ARRAY,
                 response -> castArray(handleArrayResponse(response), String.class));
     }
 
@@ -1269,18 +1626,18 @@ public class GlideClusterClient extends BaseClient
     @Override
     public CompletableFuture<String> unwatch(@NonNull Route route) {
         return commandManager.submitNewCommand(
-                UnWatch, new String[0], route, this::handleStringResponse);
+                UnWatch, EMPTY_STRING_ARRAY, route, this::handleStringResponse);
     }
 
     @Override
     public CompletableFuture<String> unwatch() {
-        return commandManager.submitNewCommand(UnWatch, new String[0], this::handleStringResponse);
+        return commandManager.submitNewCommand(UnWatch, EMPTY_STRING_ARRAY, this::handleStringResponse);
     }
 
     @Override
     public CompletableFuture<String> randomKey(@NonNull Route route) {
         return commandManager.submitNewCommand(
-                RandomKey, new String[0], route, this::handleStringOrNullResponse);
+                RandomKey, EMPTY_STRING_ARRAY, route, this::handleStringOrNullResponse);
     }
 
     @Override
@@ -1295,7 +1652,7 @@ public class GlideClusterClient extends BaseClient
     @Override
     public CompletableFuture<String> randomKey() {
         return commandManager.submitNewCommand(
-                RandomKey, new String[0], this::handleStringOrNullResponse);
+                RandomKey, EMPTY_STRING_ARRAY, this::handleStringOrNullResponse);
     }
 
     @Override